% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{$,enve.RecPlot2.Peak-method}
\alias{$,enve.RecPlot2.Peak-method}
\title{Attribute accessor}
\usage{
\S4method{$}{enve.RecPlot2.Peak}(x, name)
}
\arguments{
\item{x}{Object}

\item{name}{Attribute name}
}
\description{
Attribute accessor
}
