\name{enve.__gss}
\alias{enve.__gss}
\title{enve   gss}
\description{Internal ancilliary function (see `enve.gss`).}
\usage{enve.__gss(rep, frx, selection, dimensions, dots, dist.method, 
    summary.fx, dist)}
\arguments{
  \item{rep}{
}
  \item{frx}{
}
  \item{selection}{
}
  \item{dimensions}{
}
  \item{dots}{
}
  \item{dist.method}{
}
  \item{summary.fx}{
}
  \item{dist}{
}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





