% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoprune.R
\name{enve.__prune.iter}
\alias{enve.__prune.iter}
\title{Enveomics: Prune Iter (Internal Function)}
\usage{
enve.__prune.iter(t, dist, min_dist, quiet)
}
\arguments{
\item{t}{A \strong{phylo} object.}

\item{dist}{Cophenetic distance matrix.}

\item{min_dist}{Minimum distance.}

\item{quiet}{If running quietly.}
}
\value{
Returns a \strong{phylo} object.
}
\description{
Internal function for \code{\link{enve.prune.dist}}.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
