% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoprune.R
\name{enve.__prune.reduce}
\alias{enve.__prune.reduce}
\title{Enveomics: Prune Reduce (Internal Function)}
\usage{
enve.__prune.reduce(t, nodes, min_dist, quiet)
}
\arguments{
\item{t}{A \strong{phylo} object.}

\item{nodes}{Vector of nodes.}

\item{min_dist}{Minimum distance.}

\item{quiet}{If running quietly.}
}
\value{
A \strong{phylo} object.
}
\description{
Internal function for \code{\link{enve.prune.dist}}.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
