% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribs.R
\name{enve.__tribs}
\alias{enve.__tribs}
\title{Enveomics: TRIBS - Internal Ancillary Function}
\usage{
enve.__tribs(rep, frx, selection, dimensions, dots, dist.method,
  summary.fx, dist)
}
\arguments{
\item{rep}{Replicates}

\item{frx}{Fraction}

\item{selection}{Selection}

\item{dimensions}{Dimensions}

\item{dots}{Sampling points}

\item{dist.method}{Distance method}

\item{summary.fx}{Summary function}

\item{dist}{Distance}
}
\description{
Internal ancillary function (see \code{\link{enve.tribs}}).
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
