% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve.R
\name{enve.col2alpha}
\alias{enve.col2alpha}
\title{Enveomics: Color to Alpha}
\usage{
enve.col2alpha(x, alpha)
}
\arguments{
\item{x}{A vector of any value base colors.}

\item{alpha}{Alpha level to set (in the 0-1 range).}
}
\description{
Takes a vector of colors and sets the alpha.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
