% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2dist.R
\name{enve.df2dist.list}
\alias{enve.df2dist.list}
\title{Enveomics: Data Frame to Dist (List)}
\usage{
enve.df2dist.list(
  x,
  groups,
  obj1.index = 1,
  obj2.index = 2,
  dist.index = 3,
  empty.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{A dataframe (or coercible object) with at least three columns:
\enumerate{
   \item ID of the object 1, 
   \item ID of the object 2, and 
   \item distance between the two objects.}}

\item{groups}{Named array where the IDs correspond to the object IDs,
and the values correspond to the group.}

\item{obj1.index}{Index of the column containing the ID of the object 1.}

\item{obj2.index}{Index of the column containing the ID of the object 2.}

\item{dist.index}{Index of the column containing the distance.}

\item{empty.rm}{Remove incomplete matrices.}

\item{...}{Any other parameters supported by \code{\link{enve.df2dist}}.}
}
\value{
Returns a \strong{list} of \strong{dist} objects.
}
\description{
Transform a dataframe (or coercible object, like a table)
into a \strong{list} of \strong{dist} objects, one per group.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
