% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve.R
\docType{class}
\name{enve.GrowthCurve-class}
\alias{enve.GrowthCurve-class}
\alias{enve.GrowthCurve}
\title{Enveomics: Growth Curve S4 Class}
\description{
Enve-omics representation of fitted growth curves.
}
\section{Slots}{

\describe{
\item{\code{design}}{\code{(array)} Experimental design of the experiment.}

\item{\code{models}}{\code{(list)} Fitted growth curve models.}

\item{\code{predict}}{\code{(list)} Fitted growth curve values.}

\item{\code{call}}{\code{(call)} Call producing this object.}
}}

\author{
Luis M. Rodriguez-R [aut, cre]
}
