\name{enve.GSS.merge}
\alias{enve.GSS.merge}
\title{enve GSS merge}
\description{Merges two `enve.GSS` objects generated from the same objects at different subsampling levels.}
\usage{enve.GSS.merge(x, y)}
\arguments{
  \item{x}{First `enve.GSS` object.}
  \item{y}{Second `enve.GSS` object.}
}

\value{Returns an `enve.GSS` object.}

\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





