\name{enve.recplot2.__closestPeak}
\alias{enve.recplot2.__closestPeak}
\title{enve recplot2   closestPeak}
\description{Internal ancilliary function (see `enve.recplot2.findPeaks`).}
\usage{enve.recplot2.__closestPeak(peak, peaks)}
\arguments{
  \item{peak}{
}
  \item{peaks}{
}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





