% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.__counts}
\alias{enve.recplot2.__counts}
\title{Enveomics: Recruitment Plot (2) Internal Ancillary Function}
\usage{
enve.recplot2.__counts(x, pos.breaks, id.breaks, rec.idcol)
}
\arguments{
\item{x}{\code{\link{enve.RecPlot2}} object}

\item{pos.breaks}{Position breaks}

\item{id.breaks}{Identity breaks}

\item{rec.idcol}{Identity column to use}
}
\value{
2-dimensional matrix of counts per identity and position bins.
}
\description{
Internal ancillary function (see \code{\link{enve.recplot2}}).
}
\author{
Luis M. Rodriguez-R [aut, cre]

Kenji Gerhardt [aut]
}
