\name{enve.recplot2.__findPeak}
\alias{enve.recplot2.__findPeak}
\title{enve recplot2   findPeak}
\description{Internall ancilliary function (see `enve.recplot2.findPeaks`).}
\usage{enve.recplot2.__findPeak(lsd1, min.points, quant.est, mlv.opts, 
    fitdist.opts, with.skewness, optim.rounds, optim.epsilon, 
    n.total, merge.logdist, verbose)}
\arguments{
  \item{lsd1}{
}
  \item{min.points}{
}
  \item{quant.est}{
}
  \item{mlv.opts}{
}
  \item{fitdist.opts}{
}
  \item{with.skewness}{
}
  \item{optim.rounds}{
}
  \item{optim.epsilon}{
}
  \item{n.total}{
}
  \item{merge.logdist}{
}
  \item{verbose}{
}
}



\author{Luis M. Rodriguez-R [aut, cre]}





