\name{enve.recplot2.__findPeaks}
\alias{enve.recplot2.__findPeaks}
\title{enve recplot2   findPeaks}
\description{Internal ancilliary function (see `enve.recplot2.findPeaks`).}
\usage{enve.recplot2.__findPeaks(peaks.opts)}
\arguments{
  \item{peaks.opts}{
}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





