\name{enve.recplot2.__whichClosestPeak}
\alias{enve.recplot2.__whichClosestPeak}
\title{enve recplot2   whichClosestPeak}
\description{Internal ancilliary function (see `enve.recplot2.findPeaks`).}
\usage{enve.recplot2.__whichClosestPeak(peak, peaks)}
\arguments{
  \item{peak}{
}
  \item{peaks}{
}
}



\author{Luis M. Rodriguez-R [aut, cre]}





