\name{enve.recplot2.aucp}
\alias{enve.recplot2.aucp}
\title{enve recplot2 aucp}
\description{Estimate the Area Under the Core Peak (AUCP) from a recruitment plot
or the core peak. If a non-core peak is passed, it simply estimates the
area under that peak.}
\usage{enve.recplot2.aucp(x)}
\arguments{
  \item{x}{`enve.RecPlot2` or `enve.RecPlot2.Peak` object.}
}



\author{Luis M. Rodriguez-R [aut, cre]}





