\name{enve.recplot2.changeCutoff}
\alias{enve.recplot2.changeCutoff}
\title{enve recplot2 changeCutoff}
\description{Change the intra-species cutoff of an existing recruitment plot.}
\usage{enve.recplot2.changeCutoff(rp, new.cutoff = 98)}
\arguments{
  \item{rp}{enve.RecPlot2 object.}
  \item{new.cutoff}{New cutoff to use.}
}



\author{Luis M. Rodriguez-R [aut, cre]}





