\name{enve.recplot2.findPeaks.__em_e}
\alias{enve.recplot2.findPeaks.__em_e}
\title{enve recplot2 findPeaks   em e}
\description{Internal ancilliary function (see `enve.recplot2.findPeaks.em`).}
\usage{enve.recplot2.findPeaks.__em_e(x, theta)}
\arguments{
  \item{x}{
}
  \item{theta}{
}
}



\author{Luis M. Rodriguez-R [aut, cre]}





