% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.findPeaks.__emauto_one}
\alias{enve.recplot2.findPeaks.__emauto_one}
\title{Enveomics: Recruitment Plot (2) EMauto Peak Finder - Internal Ancillary Function}
\usage{
enve.recplot2.findPeaks.__emauto_one(x, comp, do_crit, best, verbose, ...)
}
\arguments{
\item{x}{\code{\link{enve.RecPlot2}} object.}

\item{comp}{Components.}

\item{do_crit}{Function estimating the criterion.}

\item{best}{Best solution thus far.}

\item{verbose}{If verbose.}

\item{...}{Additional parameters for \code{\link{enve.recplot2.findPeaks.em}}.}
}
\value{
Updated solution with the same structure as \code{best}.
}
\description{
Internal ancillary function (see
\code{\link{enve.recplot2.findPeaks.emauto}}).
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
