% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.findPeaks.__mow_one}
\alias{enve.recplot2.findPeaks.__mow_one}
\title{Enveomics: Recruitment Plot (2) Mowing Peak Finder - Internal Ancillary Function 1}
\usage{
enve.recplot2.findPeaks.__mow_one(
  lsd1,
  min.points,
  quant.est,
  mlv.opts,
  fitdist.opts,
  with.skewness,
  optim.rounds,
  optim.epsilon,
  n.total,
  merge.logdist,
  verbose,
  log
)
}
\arguments{
\item{lsd1}{Vector of log-transformed sequencing depths}

\item{min.points}{Minimum number of points}

\item{quant.est}{Quantile estimate}

\item{mlv.opts}{List of options for \code{mlv}}

\item{fitdist.opts}{List of options for \code{fitdist}}

\item{with.skewness}{If skewed-normal should be used}

\item{optim.rounds}{Maximum number of optimization rounds}

\item{optim.epsilon}{Minimum difference considered negligible}

\item{n.total}{Global number of windows}

\item{merge.logdist}{Attempted \code{merge.logdist} parameter}

\item{verbose}{If verbose}

\item{log}{If log-transformed depths}
}
\value{
Return an \code{enve.RecPlot2.Peak} object.
}
\description{
Internal ancillary function (see
\code{\link{enve.recplot2.findPeaks.mower}}).
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
