\name{enve.recplot2.findPeaks.__mower}
\alias{enve.recplot2.findPeaks.__mower}
\title{enve recplot2 findPeaks   mower}
\description{Internal ancilliary function (see `enve.recplot2.findPeaks.mower`).}
\usage{enve.recplot2.findPeaks.__mower(peaks.opts)}
\arguments{
  \item{peaks.opts}{
}
}



\author{Luis M. Rodriguez-R [aut, cre]}





