% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enve.selvector}
\alias{enve.selvector}
\title{Enveomics: Selection vector}
\usage{
enve.selvector(sel, dim.names)
}
\arguments{
\item{sel}{A vector of numbers, characters, or booleans.}

\item{dim.names}{A vector of names from which to select.}
}
\value{
Returns a logical vector with the same length as \code{dim.name}.
}
\description{
Normalizes a selection vector \code{sel} to a logical vector with indexes
from \code{dim.names}.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
