% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribs.R
\name{enve.TRIBS.merge}
\alias{enve.TRIBS.merge}
\title{Enveomics: TRIBS Merge}
\usage{
enve.TRIBS.merge(x, y)
}
\arguments{
\item{x}{First \code{\link{enve.TRIBS}} object.}

\item{y}{Second \code{\link{enve.TRIBS}} object.}
}
\value{
Returns an \code{\link{enve.TRIBS}} object.
}
\description{
Merges two \code{\link{enve.TRIBS}} objects generated from the same objects at
different subsampling levels.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
