\name{enveomics.R-package}
\alias{enveomics.R-package}
\alias{enveomics.R}
\docType{package}
\title{Various R functions from the Kostas lab}
\description{A collection of R functions of common use in the Kostas Lab.}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{enveomics.R}
\packageIndices{enveomics.R}
}
\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}

\keyword{ package }


