
\name{growth.curves}
\docType{data}
\alias{growth.curves}
\title{Bacterial growth curves for three Escherichia coli mutants}
\description{
  This data set provides time (first column) and three triplicated growth
  curves as optical density at 600nm (OD_600nm) for different mutants of E.
  coli.
}
\usage{growth.curves}
\format{A data frame with 16 rows (times) and 10 rows (times and OD_600nm).}
\keyword{datasets}

