\name{plot.enve.GSStest}
\alias{plot.enve.GSStest}
\title{plot enve GSStest}
\description{Plots an `enve.GSStest` object.}
\usage{\method{plot}{enve.GSStest}(x, type = c("overlap", "difference"), col = "#00000044", 
    col1 = col, col2 = "#44001144", ylab = "Probability", xlim = range(attr(x, 
        "dist.mids")), ylim = c(0, max(c(attr(x, "all.dist"), 
        attr(x, "sel.dist")))), ...)}
\arguments{
  \item{x}{
}
  \item{type}{
}
  \item{col}{
}
  \item{col1}{
}
  \item{col2}{
}
  \item{ylab}{
}
  \item{xlim}{
}
  \item{ylim}{
}
  \item{\dots}{
}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





