\name{summary.enve.GrowthCurve}
\alias{summary.enve.GrowthCurve}
\title{summary enve GrowthCurve}
\description{Summary of an `enve.GrowthCurve` object.}
\usage{\method{summary}{enve.GrowthCurve}(object, ...)}
\arguments{
  \item{object}{`enve.GrowthCurve` object.}
  \item{\dots}{No additional parameters are currently supported.}
}



\author{Luis M. Rodriguez-R [aut, cre]}





