\name{summary.enve.GSS}
\alias{summary.enve.GSS}
\title{summary enve GSS}
\description{Summary of an `enve.GSS` object.}
\usage{\method{summary}{enve.GSS}(object, ...)}
\arguments{
  \item{object}{`enve.GSS` object.}
  \item{\dots}{No additional parameters are currently supported.}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





