\name{summary.enve.GSStest}
\alias{summary.enve.GSStest}
\title{summary enve GSStest}
\description{Summary of an `enve.GSStest` object.}
\usage{\method{summary}{enve.GSStest}(object, ...)}
\arguments{
  \item{object}{`enve.GSStest` object.}
  \item{\dots}{No additional parameters are currently supported.}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





