% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribs.R
\name{summary.enve.TRIBS}
\alias{summary.enve.TRIBS}
\title{Enveomics: TRIBS Summary}
\usage{
\method{summary}{enve.TRIBS}(object, ...)
}
\arguments{
\item{object}{\code{\link{enve.TRIBS}} object.}

\item{...}{No additional parameters are currently supported.}
}
\value{
No return value.
}
\description{
Summary of an \code{\link{enve.TRIBS}} object.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
