% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribs.R
\name{summary.enve.TRIBStest}
\alias{summary.enve.TRIBStest}
\title{Enveomics: TRIBS Summary Test}
\usage{
\method{summary}{enve.TRIBStest}(object, ...)
}
\arguments{
\item{object}{\code{\link{enve.TRIBStest}} object.}

\item{...}{No additional parameters are currently supported.}
}
\description{
Summary of an \code{\link{enve.TRIBStest}} object.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
