% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve.R
\name{$,enve.GrowthCurve-method}
\alias{$,enve.GrowthCurve-method}
\title{Attribute accessor}
\usage{
\S4method{$}{enve.GrowthCurve}(x, name)
}
\arguments{
\item{x}{Object}

\item{name}{Attribute name}
}
\description{
Attribute accessor
}
