\name{enve.__prune.iter}
\alias{enve.__prune.iter}
\title{enve   prune iter}
\description{Internal function for enve.prune.dist}
\usage{enve.__prune.iter(t, dist, min_dist, quiet)}
\arguments{
  \item{t}{
}
  \item{dist}{
}
  \item{min_dist}{
}
  \item{quiet}{
}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





