\name{enve.__prune.reduce}
\alias{enve.__prune.reduce}
\title{enve   prune reduce}
\description{Internal function for enve.prune.dist}
\usage{enve.__prune.reduce(t, nodes, min_dist, quiet)}
\arguments{
  \item{t}{
}
  \item{nodes}{
}
  \item{min_dist}{
}
  \item{quiet}{
}
}



\author{Luis M. Rodriguez-R [aut, cre]}





