\name{enve.__tribs}
\alias{enve.__tribs}
\title{enve   tribs}
\description{Internal ancilliary function (see `enve.tribs`).}
\usage{enve.__tribs(rep, frx, selection, dimensions, dots, dist.method, 
    summary.fx, dist)}
\arguments{
  \item{rep}{
}
  \item{frx}{
}
  \item{selection}{
}
  \item{dimensions}{
}
  \item{dots}{
}
  \item{dist.method}{
}
  \item{summary.fx}{
}
  \item{dist}{
}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





