\name{enve.col.alpha}
\alias{enve.col.alpha}
\title{enve col alpha}
\description{Modify alpha in a color (or vector of colors).}
\usage{enve.col.alpha(col, alpha = 1/2)}
\arguments{
  \item{col}{Color or vector of colors. It can be any value supported by `col2rgb`, 
such as 'darkred' or '#009988'.}
  \item{alpha}{Alpha value to add to the color, from 0 to 1.}
}

\value{Returns a color or a vector of colors in hex notation including alpha.}

\author{Luis M. Rodriguez-R [aut, cre]}





