% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enve.col.alpha}
\alias{enve.col.alpha}
\title{Enveomics: Color Alpha}
\usage{
enve.col.alpha(col, alpha = 1/2)
}
\arguments{
\item{col}{Color or vector of colors. It can be any value supported by 
\code{\link[grDevices]{col2rgb}}, such as \code{darkred} or \code{#009988}.}

\item{alpha}{Alpha value to add to the color, from 0 to 1.}
}
\value{
Returns a color or a vector of colors in \emph{hex} notation, 
including \code{alpha}.
}
\description{
Modify alpha in a color (or vector of colors).
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
