\name{enve.col2alpha}
\alias{enve.col2alpha}
\title{enve col2alpha}
\description{Takes a vector of colors and sets the alpha.}
\usage{enve.col2alpha(x, alpha)}
\arguments{
  \item{x}{A vector of any value base colors.}
  \item{alpha}{Alpha level to set (in the 0-1 range).}
}



\author{Luis M. Rodriguez-R [aut, cre]}





