\name{enve.recplot2.__counts}
\alias{enve.recplot2.__counts}
\title{enve recplot2   counts}
\description{Internal ancilliary function (see `enve.gss`).}
\usage{enve.recplot2.__counts(x, pos.breaks, id.breaks, rec.idcol)}
\arguments{
  \item{x}{
}
  \item{pos.breaks}{
}
  \item{id.breaks}{
}
  \item{rec.idcol}{
}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





