\name{enve.recplot2.__peakHist}
\alias{enve.recplot2.__peakHist}
\title{enve recplot2   peakHist}
\description{Internal ancilliary function (see `enve.RecPlot2.Peak`).}
\usage{enve.recplot2.__peakHist(x, mids, counts = TRUE)}
\arguments{
  \item{x}{
}
  \item{mids}{
}
  \item{counts}{
}
}



\author{Luis M. Rodriguez-R [aut, cre]}





