% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.__peakHist}
\alias{enve.recplot2.__peakHist}
\title{Enveomics: Recruitment Plot (2) Peak S4 Class - Internal Ancillary Function}
\usage{
enve.recplot2.__peakHist(x, mids, counts = TRUE)
}
\arguments{
\item{x}{\code{\link{enve.RecPlot2.Peak}} object}

\item{mids}{Midpoints}

\item{counts}{Counts}
}
\value{
A numeric vector of counts (histogram)
}
\description{
Internal ancillary function (see \code{\link{enve.RecPlot2.Peak}}).
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
