% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.__whichClosestPeak}
\alias{enve.recplot2.__whichClosestPeak}
\title{Enveomics: Recruitment Plot (2) Peak Finder - Internal Ancillary Function}
\usage{
enve.recplot2.__whichClosestPeak(peak, peaks)
}
\arguments{
\item{peak}{Query \code{\link{enve.RecPlot2.Peak}} object}

\item{peaks}{list of \code{\link{enve.RecPlot2.Peak}} objects}
}
\value{
A numeric index out of \code{peaks}.
}
\description{
Internal ancillary function (see \code{\link{enve.recplot2.findPeaks}}).
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
