% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.changeCutoff}
\alias{enve.recplot2.changeCutoff}
\title{Enveomics: Recruitment Plot (2) Change Cutoff}
\usage{
enve.recplot2.changeCutoff(rp, new.cutoff = 98)
}
\arguments{
\item{rp}{\code{\link{enve.RecPlot2}} object.}

\item{new.cutoff}{New cutoff to use.}
}
\value{
The modified \code{\link{enve.RecPlot2}} object.
}
\description{
Change the intra-species cutoff of an existing recruitment plot.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
