\name{enve.recplot2.corePeak}
\alias{enve.recplot2.corePeak}
\title{enve recplot2 corePeak}
\description{Finds the peak in a list of peaks that is most likely to represent the "core genome" of a population.}
\usage{enve.recplot2.corePeak(x)}
\arguments{
  \item{x}{`list` of `enve.recplot2.peak` objects.}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





