% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.corePeak}
\alias{enve.recplot2.corePeak}
\title{Enveomics: Recruitment Plot (2) Core Peak Finder}
\usage{
enve.recplot2.corePeak(x)
}
\arguments{
\item{x}{\code{list} of \code{\link{enve.RecPlot2.Peak}} objects.}
}
\description{
Finds the peak in a list of peaks that is most likely to represent the
"core genome" of a population.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
