% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.findPeaks.__em_e}
\alias{enve.recplot2.findPeaks.__em_e}
\title{Enveomics: Recruitment Plot (2) EM Peak Finder - Internal Ancillary Function Expectation}
\usage{
enve.recplot2.findPeaks.__em_e(x, theta)
}
\arguments{
\item{x}{Vector of log-transformed sequencing depths}

\item{theta}{Parameters list}
}
\description{
Internal ancillary function (see \code{\link{enve.recplot2.findPeaks.em}}).
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
