\name{enve.recplot2.findPeaks.__em_m}
\alias{enve.recplot2.findPeaks.__em_m}
\title{enve recplot2 findPeaks   em m}
\description{Internal ancilliary function (see `enve.recplot2.findPeaks.em`}
\usage{enve.recplot2.findPeaks.__em_m(x, posterior)}
\arguments{
  \item{x}{
}
  \item{posterior}{
}
}



\author{Luis M. Rodriguez-R [aut, cre]}





