\name{enve.recplot2.findPeaks.__emauto_one}
\alias{enve.recplot2.findPeaks.__emauto_one}
\title{enve recplot2 findPeaks   emauto one}
\description{Internal ancilliary function (see `enve.recplot2.findPeaks.emauto).}
\usage{enve.recplot2.findPeaks.__emauto_one(x, comp, do_crit, best, 
    verbose, ...)}
\arguments{
  \item{x}{
}
  \item{comp}{
}
  \item{do_crit}{
}
  \item{best}{
}
  \item{verbose}{
}
  \item{\dots}{
}
}



\author{Luis M. Rodriguez-R [aut, cre]}





