\name{enve.recplot2.findPeaks.__mow_one}
\alias{enve.recplot2.findPeaks.__mow_one}
\title{enve recplot2 findPeaks   mow one}
\description{Internall ancilliary function (see `enve.recplot2.findPeaks.mower`).}
\usage{enve.recplot2.findPeaks.__mow_one(lsd1, min.points, quant.est, 
    mlv.opts, fitdist.opts, with.skewness, optim.rounds, optim.epsilon, 
    n.total, merge.logdist, verbose, log)}
\arguments{
  \item{lsd1}{
}
  \item{min.points}{
}
  \item{quant.est}{
}
  \item{mlv.opts}{
}
  \item{fitdist.opts}{
}
  \item{with.skewness}{
}
  \item{optim.rounds}{
}
  \item{optim.epsilon}{
}
  \item{n.total}{
}
  \item{merge.logdist}{
}
  \item{verbose}{
}
  \item{log}{
}
}



\author{Luis M. Rodriguez-R [aut, cre]}





