% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.findPeaks.__mower}
\alias{enve.recplot2.findPeaks.__mower}
\title{Enveomics: Recruitment Plot (2) Mowing Peak Finder - Internal Ancillary Function 2}
\usage{
enve.recplot2.findPeaks.__mower(peaks.opts)
}
\arguments{
\item{peaks.opts}{List of options for \code{\link{enve.recplot2.findPeaks.__mow_one}}}
}
\description{
Internal ancillary function (see \code{\link{enve.recplot2.findPeaks.mower}}).
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
