% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recplot2.R
\name{enve.recplot2.findPeaks}
\alias{enve.recplot2.findPeaks}
\title{Enveomics: Recruitment Plot (2) Peak Finder}
\usage{
enve.recplot2.findPeaks(x, method = "emauto", ...)
}
\arguments{
\item{x}{An \code{\link{enve.RecPlot2}} object.}

\item{method}{Peak-finder method. This should be one of:
\itemize{
   \item \strong{emauto}
   (Expectation-Maximization with auto-selection of components)
   \item \strong{em}
   (Expectation-Maximization)
   \item \strong{mower}
   (Custom distribution-mowing method)
}}

\item{...}{Any additional parameters supported by
\code{\link{enve.recplot2.findPeaks}}.}
}
\value{
Returns a list of \code{\link{enve.RecPlot2.Peak}} objects.
}
\description{
Identifies peaks in the population histogram potentially indicating
sub-population mixtures.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
