\name{enve.TRIBS.merge}
\alias{enve.TRIBS.merge}
\title{enve TRIBS merge}
\description{Merges two `enve.TRIBS` objects generated from the same objects at
different subsampling levels.}
\usage{enve.TRIBS.merge(x, y)}
\arguments{
  \item{x}{First `enve.TRIBS` object.}
  \item{y}{Second `enve.TRIBS` object.}
}

\value{Returns an `enve.TRIBS` object.}

\author{Luis M. Rodriguez-R [aut, cre]}





