% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribs.R
\name{enve.tribs.test}
\alias{enve.tribs.test}
\title{Enveomics: TRIBS Test}
\usage{
enve.tribs.test(dist, selection, bins = 50, ...)
}
\arguments{
\item{dist}{Distances as \code{dist} object.}

\item{selection}{Selection defining the subset.}

\item{bins}{Number of bins to evaluate in the range of distances.}

\item{...}{Any other parameters supported by \code{\link{enve.tribs}}, 
except \code{subsamples}.}
}
\value{
Returns an \code{\link{enve.TRIBStest}} object.
}
\description{
Estimates the empirical difference between all the distances in a set of
objects and a subset, together with its statistical significance.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
