% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enve.truncate}
\alias{enve.truncate}
\title{Enveomics: Truncate}
\usage{
enve.truncate(x, f = 0.95, FUN = mean)
}
\arguments{
\item{x}{A vector of numbers.}

\item{f}{The fraction of values to retain.}

\item{FUN}{Summary function to apply to the vectors. To obtain the 
truncated vector itself, use \code{c}.}
}
\value{
Returns the summary \code{(FUN)} of the truncated vector.
}
\description{
Removes the \code{n} highest and lowest values from a vector, and applies
summary function. The value of \code{n} is determined such that the central
range is used, corresponding to the \code{f} fraction of values.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
