\name{enveomics.R-package}
\alias{enveomics.R-package}
\alias{enveomics.R}
\docType{package}
\title{Various R Functions from the Kostas Lab}
\description{A collection of R functions of common use in
the Kostas Lab for microbial ecology.}
\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}

\keyword{ package }



