\name{enveomics.R-package}
\alias{enveomics.R-package}
\alias{enveomics.R}
\docType{package}
\title{Various Utilities for Microbial Genomics and Metagenomics}
\description{A collection of functions for microbial ecology and other
applications of genomics and metagenomics. Companion package for the
Enveomics Collection (Rodriguez-R, L.M. and Konstantinidis, K.T., 2016
<DOI:10.7287/peerj.preprints.1900v1>).}
\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}

\keyword{ package }



