\name{enveomics.R-package}
\alias{enveomics.R-package}
\alias{enveomics.R}
\docType{package}
\title{Various R functions from the Kostas lab}
\description{A collection of R functions of common use in the
Kostas Lab (http://enve-omics.gatech.edu/).}
\details{
\tabular{ll}{Package: \tab enveomics.R\cr
Title: \tab Various R functions from the Kostas lab\cr
Type: \tab Package\cr
Version: \tab 2013.12.10\cr
Authors@R: \tab c(person("Luis M.","Rodriguez-R",role=c("aut","cre"),
email="lmrodriguezr@gmail.com"))\cr
URL: \tab http://enve-omics.gatech.edu/\cr
Depends: \tab R (>= 2.9)\cr
Suggests: \tab stats, gplots, modeest, fitdistrplus, moments\cr
License: \tab Artistic-2\cr
LazyLoad: \tab yes\cr
Encoding: \tab UTF-8\cr
Author: \tab Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]\cr
Maintainer: \tab Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]\cr}
}
\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}

\keyword{ package }


