% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribs.R
\name{plot.enve.TRIBS}
\alias{plot.enve.TRIBS}
\title{Enveomics: TRIBS Plot}
\usage{
\method{plot}{enve.TRIBS}(x, new = TRUE, type = c("boxplot", "points"),
  col = "#00000044", pt.cex = 1/2, pt.pch = 19, pt.col = col,
  ln.col = col, ...)
}
\arguments{
\item{x}{\code{\link{enve.TRIBS}} object to plot.}

\item{new}{Should a new canvas be drawn?}

\item{type}{Type of plot. The \strong{points} plot shows all the replicates, the 
\strong{boxplot} plot represents the values found by
\code{\link[grDevices]{boxplot.stats}}.
as areas, and plots the outliers as points.}

\item{col}{Color of the areas and/or the points.}

\item{pt.cex}{Size of the points.}

\item{pt.pch}{Points character.}

\item{pt.col}{Color of the points.}

\item{ln.col}{Color of the lines.}

\item{...}{Any additional parameters supported by \code{plot}.}
}
\description{
Plot an \code{\link{enve.TRIBS}} object.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
