% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribs.R
\name{plot.enve.TRIBStest}
\alias{plot.enve.TRIBStest}
\title{Enveomics: TRIBS Plot Test}
\usage{
\method{plot}{enve.TRIBStest}(x, type = c("overlap", "difference"),
  col = "#00000044", col1 = col, col2 = "#44001144",
  ylab = "Probability", xlim = range(attr(x, "dist.mids")),
  ylim = c(0, max(c(attr(x, "all.dist"), attr(x, "sel.dist")))), ...)
}
\arguments{
\item{x}{\code{\link{enve.TRIBStest}} object to plot.}

\item{type}{What to plot. \code{overlap} generates a plot of the two contrasting empirical
PDFs (to compare against each other), \code{difference} produces a plot of the
differences between the empirical PDFs (to compare against zero).}

\item{col}{Main color of the plot if type=\code{difference}.}

\item{col1}{First color of the plot if type=\code{overlap}.}

\item{col2}{Second color of the plot if type=\code{overlap}.}

\item{ylab}{Y-axis label.}

\item{xlim}{X-axis limits.}

\item{ylim}{Y-axis limits.}

\item{...}{Any other graphical arguments.}
}
\description{
Plots an \code{\link{enve.TRIBStest}} object.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
