% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve.R
\name{summary.enve.GrowthCurve}
\alias{summary.enve.GrowthCurve}
\title{Enveomics: Summary of Growth Curve}
\usage{
\method{summary}{enve.GrowthCurve}(object, ...)
}
\arguments{
\item{object}{An \code{\link{enve.GrowthCurve}} object.}

\item{...}{No additional parameters are currently supported.}
}
\value{
No return value.
}
\description{
Summary of an \code{\link{enve.GrowthCurve}} object.
}
\author{
Luis M. Rodriguez-R [aut, cre]
}
