\name{summary.enve.TRIBS}
\alias{summary.enve.TRIBS}
\title{summary enve TRIBS}
\description{Summary of an `enve.TRIBS` object.}
\usage{\method{summary}{enve.TRIBS}(object, ...)}
\arguments{
  \item{object}{`enve.TRIBS` object.}
  \item{\dots}{No additional parameters are currently supported.}
}



\author{Luis M. Rodriguez-R <lmrodriguezr@gmail.com> [aut, cre]}





