\name{summary.enve.TRIBStest}
\alias{summary.enve.TRIBStest}
\title{summary enve TRIBStest}
\description{Summary of an `enve.TRIBStest` object.}
\usage{\method{summary}{enve.TRIBStest}(object, ...)}
\arguments{
  \item{object}{`enve.TRIBStest` object.}
  \item{\dots}{No additional parameters are currently supported.}
}



\author{Luis M. Rodriguez-R [aut, cre]}





